tmux: make tmux navigation shorter and smarter
The new `tmux-navigate` script detects that direction and the position
of the pane to determine if it should switch pane or window. By default,
the ALT+hjl will try to switch between panes. If the pane is at the edge
of the switch direction than it will try to switch window. This is the
similar switching to zellij.